backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.7.4
structlog==24.2.0
python-multipart==0.0.9
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
textblob>=0.18.0
pytrends>=4.9.2
feedparser>=6.0.10
groq>=0.9.0
python-dotenv>=1.0.0
